finished writing break_tie

+++ Meeting/views/break_tie.py
@@ -12,13 +12,26 @@ def break_tie(request, meeting_id, vote_id):
     if vote.token_set.meeting == meeting and vote.state == vote.READY:
         return JsonResponse({'error': 'meeting vote mismatch'}, status=401)
 
-    winner = request.POST['winner_id']
-    if not Tie.objects.filter(vote=vote, option_id=winner).exists():
-        return JsonResponse({'error': 'winner_id was not an option'}, status=401)
+    if vote.state == Vote.NEEDS_TIE_BREAKER:
+        if request.method == "GET":
+            options = {}
+            for tied_candidate in vote.tie_set:
+                options[tied_candidate.option.id] = {
+                    "name": tied_candidate.option.name
+                }
+            message = {"type": "tied_options",
+                       "options": options,
+                       }
+            return JsonResponse(message)
+        elif request.method == "POST":
+            winner = request.POST['winner_id']
+            if not Tie.objects.filter(vote=vote, option_id=winner).exists():
+                return JsonResponse({'error': 'winner_id was not an option'}, status=401)
 
-    for tie in Tie.objects.filter(vote=vote).exclude(option_id=winner).all():
-        tie.delete()
-    vote.state = vote.COUNTING
-    vote.save()
-    message = {'type': 'success'}
-    return JsonResponse(message)
+            for tie in Tie.objects.filter(vote=vote).exclude(option_id=winner).all():
+                tie.delete()
+            vote.state = vote.COUNTING
+            vote.save()
+            return JsonResponse({'type': 'success'})
+    else:
+        return JsonResponse({"error": "designated stalemate associate un needed"})
